% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_model.R
\name{apply_operations}
\alias{apply_operations}
\title{Aplicar una lista de operaciones a un data.frame}
\usage{
apply_operations(df, ops)
}
\arguments{
\item{df}{Data.frame base}

\item{ops}{Lista de operaciones (output de op_* functions)}
}
\value{
Data.frame transformado
}
\description{
Aplicar una lista de operaciones a un data.frame
}
