% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_path}
\alias{config_path}
\title{Configuración persistente de RVisual}
\usage{
config_path()
}
\description{
Guarda/lee configuración en un archivo JSON local
usando el directorio de datos del usuario (rappdirs-style).
}
