% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{history_log}
\alias{history_log}
\title{Registrar evento en el historial de sesión}
\usage{
history_log(history, type, detail = list())
}
\arguments{
\item{history}{reactiveVal que contiene la lista de eventos}

\item{type}{String identificador del tipo de evento}

\item{detail}{Lista con detalles del evento}
}
\description{
Registrar evento en el historial de sesión
}
