% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin.R
\name{launch_rvisual}
\alias{launch_rvisual}
\title{Lanzar RVisual}
\usage{
launch_rvisual(browser = FALSE)
}
\arguments{
\item{browser}{Lógico. Si \code{TRUE} abre en el browser del sistema en lugar del panel Viewer de RStudio. Default \code{FALSE}.}
}
\description{
Punto de entrada del addin de RStudio.
}
