% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_discovery.R
\name{load_file}
\alias{load_file}
\title{Cargar un archivo de datos según su extensión}
\usage{
load_file(path, extension)
}
\arguments{
\item{path}{Ruta al archivo}

\item{extension}{Extensión sin punto: "csv", "xlsx", "rds", "sav"}
}
\value{
Data.frame o tibble
}
\description{
Cargar un archivo de datos según su extensión
}
