% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_builder.R
\name{mod_builder_ui}
\alias{mod_builder_ui}
\title{Módulo: Constructor Visual de Operaciones}
\usage{
mod_builder_ui(id)
}
\arguments{
\item{id}{ID del módulo Shiny (string). Usado internamente por \code{NS()} para crear el namespace del módulo.}
}
\description{
Módulo: Constructor Visual de Operaciones
}
