% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_config.R
\name{mod_config_ui}
\alias{mod_config_ui}
\title{Módulo: Configuración}
\usage{
mod_config_ui(id)
}
\arguments{
\item{id}{ID del módulo Shiny (string). Usado internamente por \code{NS()} para crear el namespace del módulo.}
}
\description{
Pantalla de configuración del addin.
Permite gestionar proveedores IA, API keys, modelos y preferencias.
}
