% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{mod_history_ui_inline}
\alias{mod_history_ui_inline}
\title{Renderizar historial como UI inline (para modal)}
\usage{
mod_history_ui_inline(history_list)
}
\arguments{
\item{history_list}{Lista de eventos del historial de sesión, generada por \code{history_log()}.}
}
\description{
Renderizar historial como UI inline (para modal)
}
