% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_model.R
\name{op_arrange}
\alias{op_arrange}
\title{Crear operación: arrange}
\usage{
op_arrange(col, desc = FALSE)
}
\arguments{
\item{col}{Nombre de columna}

\item{desc}{Lógico: TRUE = descendente}
}
\description{
Crear operación: arrange
}
