% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_model.R
\name{op_filter}
\alias{op_filter}
\title{Crear operación: filter}
\usage{
op_filter(col, op, value)
}
\arguments{
\item{col}{Nombre de columna}

\item{op}{Operador: "==", "!=", ">", "<", ">=", "<="}

\item{value}{Valor de comparación}
}
\description{
Crear operación: filter
}
