% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_model.R
\name{op_group_summarise}
\alias{op_group_summarise}
\title{Crear operación: group_summarise}
\usage{
op_group_summarise(group_cols, summary_fns)
}
\arguments{
\item{group_cols}{Character vector de columnas de agrupación}

\item{summary_fns}{Lista nombrada: nombre_nueva_col = list(fn, col_origen)}
}
\description{
Crear operación: group_summarise
}
