% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_model.R
\name{op_join}
\alias{op_join}
\title{Crear operación: join}
\usage{
op_join(right_df_name, by_cols, join_type = "left")
}
\arguments{
\item{right_df_name}{Nombre del segundo dataset}

\item{by_cols}{Vector de columnas clave (mismas en ambos)}

\item{join_type}{"left", "right", "inner", "full"}
}
\description{
Crear operación: join
}
