% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_model.R
\name{op_mutate}
\alias{op_mutate}
\title{Crear operación: mutate (nueva variable)}
\usage{
op_mutate(new_col, expression)
}
\arguments{
\item{new_col}{Nombre de la nueva columna}

\item{expression}{Expresión como string (será validada antes de generar código)}
}
\description{
Crear operación: mutate (nueva variable)
}
