% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_model.R
\name{op_recode}
\alias{op_recode}
\title{Crear operación: recode}
\usage{
op_recode(col, mapping)
}
\arguments{
\item{col}{Columna a recodificar}

\item{mapping}{Lista nombrada: valor_original = valor_nuevo}
}
\description{
Crear operación: recode
}
