% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_model.R
\name{op_rename}
\alias{op_rename}
\title{Crear operación: rename}
\usage{
op_rename(old_name, new_name)
}
\arguments{
\item{old_name}{Nombre actual de la columna}

\item{new_name}{Nuevo nombre}
}
\description{
Crear operación: rename
}
