% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio_bridge.R
\name{rstudio_active_script_path}
\alias{rstudio_active_script_path}
\title{Obtener el path del script activo en RStudio}
\usage{
rstudio_active_script_path()
}
\description{
Obtener el path del script activo en RStudio
}
