% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{rvisual_ui}
\alias{rvisual_ui}
\title{UI principal de RVisual}
\usage{
rvisual_ui()
}
\value{
Un objeto shiny.tag con la UI completa
}
\description{
Define la interfaz Shiny completa del addin.
Estructura en 6 paneles navegables via sidebar.
}
