% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_discovery.R
\name{classify_visual_type}
\alias{classify_visual_type}
\title{Clasificar tipo visual de una columna (para mostrar al usuario)}
\usage{
classify_visual_type(col)
}
\arguments{
\item{col}{Vector de una columna del data.frame a clasificar.}
}
\description{
Clasificar tipo visual de una columna (para mostrar al usuario)
}
