% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio_bridge.R
\name{rstudio_run_in_console}
\alias{rstudio_run_in_console}
\title{Ejecutar código en la consola de RStudio}
\usage{
rstudio_run_in_console(code)
}
\arguments{
\item{code}{String con el código R a ejecutar en la consola de RStudio.}
}
\description{
Ejecutar código en la consola de RStudio
}
